+++ ocdsextensionregistry/extension.py
@@ -6,3 +6,11 @@ class Extension:
         self.id = data['Id']
         self.category = data['Category']
         self.core = data['Core'] == 'true'
+
+    def as_dict(self):
+        """
+        Returns the object's properties as a dictionary.
+
+        This method is defined to match the method in `ExtensionVersion`.
+        """
+        return self.__dict__