+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
+## 0.0.3 (2018-06-21)
+
+* Add `remote(basename)` method to `ExtensionVersion`, to return the contents of a file within the extension.
+* Add `as_dict()` method to `Extension` and `ExtensionVersion`, to avoid returning private properties.
+
 ## 0.0.2 (2018-06-12)
 
-* Make `ExtensionRegistry` iterable.
+* Add `get(**kwargs)` method to `ExtensionRegistry`, to get a specific extension version.
+* Make `ExtensionRegistry` iterable, to iterate over all extension versions.
 * Remove `all()` method from `ExtensionRegistry`.
-* Add `get(**kwargs)` method to `ExtensionRegistry`.
 * Add package-specific exceptions.
 
 ## 0.0.1 (2018-06-11)